creating data through post request works
