+++ db/entries.py
@@ -1,64 +1,44 @@
 from sqlalchemy.orm import Session
 
 from entries.models import Entry
-from entries.schemas import Entry
+from entries.schemas import EntryTable, DirectorTable, ActorTable, CountryTable, GenreTable
 
 
 def get_single_entry(db: Session, show_id: str):
-    return db.query(Entry).filter(Entry.show_id == show_id).first()
+    return db.query(EntryTable).filter(EntryTable.show_id == show_id).first()
 
 
 def get_entries(db: Session):
-    return db.query(Entry).all()
+    return db.query(EntryTable).all()
 
 
 def create_entry(db: Session, entry: Entry):
-    db_entry = Entry(
+    db_entry = EntryTable(
         show_id=entry.show_id,
         title=entry.title,
-        directors=[director for director in entry.directors],
-        casts=[cast for cast in entry.casts],
-        countries=[country for country in entry.countries],
         type=entry.type,
         date_added=entry.date_added,
         release_year=entry.release_year,
         rating=entry.rating,
         duration=entry.duration,
-        listed_in=[genre for genre in entry.genres],
         description=entry.description
     )
 
-    # for director in entry.directors:
-    #     try:
-    #         db_entry.directors.append(
-    #             Director(name=director.name)
-    #         )
-    #     except ValidationError:
-    #         pass
-    #
-    # for cast in entry.casts:
-    #     try:
-    #         db_entry.casts.append(
-    #             Actor(name=cast.name)
-    #         )
-    #     except ValidationError:
-    #         pass
-    #
-    # for country in entry.countries:
-    #     try:
-    #         db_entry.countries.append(
-    #             Country(name=country.name)
-    #         )
-    #     except ValidationError:
-    #         pass
-    #
-    # for genre in entry.genres:
-    #     try:
-    #         db_entry.genres.append(
-    #             Genre(name=genre.name)
-    #         )
-    #     except ValidationError:
-    #         pass
+    for director in entry.directors:
+        if not director_exists(db, director.name):
+            db_entry.directors.append(DirectorTable(name=director.name))
+
+    for cast in entry.casts:
+        if not cast_exists(db, cast.name):
+            db_entry.casts.append(ActorTable(name=cast.name))
+
+    for country in entry.countries:
+        if not country_exists(db, country.name):
+            db_entry.countries.append(CountryTable(name=country.name))
+
+    for genre in entry.listed_in:
+        if not genre_exists(db, genre.name):
+            db_entry.listed_in.append(GenreTable(name=genre.name))
 
     db.add(db_entry)
     db.commit()
@@ -67,4 +47,21 @@ def create_entry(db: Session, entry: Entry):
 
 
 def delete_entry(db: Session, show_id: str):
-    return db.query(Entry).filter(Entry.show_id == show_id).delete(synchronize_session="fetch")
+    return db.query(EntryTable).filter(EntryTable.show_id == show_id).delete(synchronize_session="fetch")
+
+
+def director_exists(db: Session, name: str):
+    return False if not db.query(DirectorTable).filter(DirectorTable.name == name).first() else True
+
+
+def cast_exists(db: Session, name: str):
+    return False if not db.query(ActorTable).filter(ActorTable.name == name).first() else True
+
+
+def country_exists(db: Session, name: str):
+    return False if not db.query(CountryTable).filter(CountryTable.name == name).first() else True
+
+
+def genre_exists(db: Session, name: str):
+    return False if not db.query(GenreTable).filter(GenreTable.name == name).first() else True
+
