+++ entries/routes.py
@@ -3,7 +3,7 @@ from typing import List
 from sqlalchemy.orm import Session
 
 from db import get_db, entries as db_entries
-from .schemas import Entry
+from .models import Entry
 
 
 router = APIRouter()
@@ -36,8 +36,8 @@ def get_entries(db: Session = Depends(get_db)):
 
 
 @router.delete("/browse/{show_id}", response_class=Response, responses={
-    200: {"message": "Successfully Deleted"},
-    404: {"message": "Entry not found."}
+    200: {"detail": "Successfully Deleted"},
+    404: {"detail": "Entry not found."}
 })
 def delete_entry(show_id: str, db: Session = Depends(get_db)):
     return Response(status_code=200) if db_entries.delete_entry(db, show_id=show_id) else Response(status_code=404)
